+++ apps/docs/.storybook/__mocks__/next/image.tsx
@@ -1,0 +1,5 @@
+import React from 'react'
+
+export default function (props) {
+  return <img {...props} />
+}
