+++ apps/docs/.storybook/__mocks__/next/link.tsx
@@ -1,0 +1,5 @@
+import React from 'react'
+
+export default function ({ children }) {
+  return <a>{children}</a>
+}
