+++ apps/docs/.storybook/__mocks__/next/router.tsx
@@ -1,0 +1,10 @@
+import React from 'react'
+
+export const useRouter = () => ({
+  route: '/',
+  pathname: '',
+  query: '',
+  asPath: '',
+  prefetch: () => {},
+  push: () => {},
+})
