+++ apps/docs/package.json
@@ -33,9 +33,9 @@
     "sass": "^1.63.3",
     "shared": "workspace:*",
     "storybook": "7.0.27",
-    "web": "workspace:*",
     "tailwind-config": "workspace:*",
     "tsconfig": "workspace:*",
-    "typescript": "^4.5.3"
+    "typescript": "^4.5.3",
+    "web": "workspace:*"
   }
 }
