+++ apps/web/index.tsx
@@ -1,0 +1,5 @@
+// storybook용 export
+
+import Navigation from '@/common/components/Navigation'
+
+export { Navigation }
