+++ apps/web/package.json
@@ -2,6 +2,8 @@
   "name": "web",
   "version": "1.0.0",
   "private": true,
+  "main": "./index.tsx",
+
   "scripts": {
     "dev": "next dev",
     "build": "next build",
