+++ packages/myll-ui/index.tsx
@@ -1,8 +1,7 @@
 import Button from './src/Components/Button/Button'
-import FooterNav from './src/Components/FooterNav'
 import Input from './src/Components/Input'
 
 export * from './src/Components/Card'
 export * from './src/Header'
 
-export { Button, FooterNav, Input }
+export { Button, Input }
